Fix Workflows ci.

+++ src/debug/views/detail.php
@@ -24,7 +24,7 @@ echo GridView::widget([
                 $timeInSeconds = $data['timestamp'] / 1000;
                 $millisecondsDiff = (int) (($timeInSeconds - floor($timeInSeconds)) * 1000);
 
-                return date('H:i:s.', floor($timeInSeconds)) . sprintf('%03d', $millisecondsDiff);
+                return date('H:i:s.', (int) floor($timeInSeconds)) . sprintf('%03d', $millisecondsDiff);
             },
             'headerOptions' => [
                 'class' => 'sort-numerical'
